Revert "added running with left shift key. fixed conversion warnings"
This reverts commit 8e593f421b9f8356966d1353363869de0524e448.

+++ Map.cpp
@@ -26,7 +26,7 @@ void Map::LoadMap(const char* LayerOneFileName, const char* LayerTwoFilename, co
 				{
 					convertedData = stoi(data);
 					MapTile MT;
-					MT.CreateMapTile(static_cast<float>(Col), static_cast<float>(Row), (MapTileType)convertedData, 32, 2, MTA, true);
+					MT.CreateMapTile(Col, Row, (MapTileType)convertedData, 32, 2, MTA, true);
 					LayerOne.push_back(MT);
 					data = "";
 					Col += 64;
